#112 where needed every api call has secur check

+++ backend/vurrukkulluk/src/main/java/com/lydia/vurrukkulluk/controller/ArticleController.java
@@ -5,6 +5,7 @@ import com.lydia.vurrukkulluk.dto.UserDto;
 import com.lydia.vurrukkulluk.model.Article;
 import com.lydia.vurrukkulluk.model.User;
 import com.lydia.vurrukkulluk.service.ArticleService;
+import com.lydia.vurrukkulluk.util.SecurityUtil;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -22,6 +23,9 @@ public class ArticleController {
 
     @Autowired
     private ModelMapper modelMapper;
+
+    @Autowired
+    private SecurityUtil securityUtil;
 
     public ArticleController(){
 
@@ -46,14 +50,23 @@ public class ArticleController {
 
     @PatchMapping("/{id}")
     public String update(@RequestBody Article article){
+        if (!securityUtil.isAdmin()) {
+            return "not authorized";
+        }
         articleService.updateArticle(article);
         return "article updated";
+
     }
 
     @DeleteMapping("/{id}")
     public String delete(@PathVariable int id){
+
+        if (!securityUtil.isAdmin()){
+            return "not authorized";
+        }
         articleService.deleteArticleById(id);
         return "article updated";
+
     }
 
     public ArticleDto convertArticleToDto(Article article){
