+++ backend/vurrukkulluk/src/main/java/com/lydia/vurrukkulluk/controller/CalendarItemController.java
@@ -2,6 +2,7 @@ package com.lydia.vurrukkulluk.controller;
 
 import com.lydia.vurrukkulluk.model.CalendarItem;
 import com.lydia.vurrukkulluk.service.CalendarItemService;
+import com.lydia.vurrukkulluk.util.SecurityUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,7 +14,9 @@ import java.util.List;
 public class CalendarItemController {
 
     @Autowired
-    CalendarItemService calendarItemService;
+    private CalendarItemService calendarItemService;
+    @Autowired
+    private SecurityUtil securityUtil;
 
     @GetMapping()
     public List<CalendarItem> getAllItems(){
@@ -22,18 +25,30 @@ public class CalendarItemController {
 
     @PostMapping()
     public String saveCalendarItem(@RequestBody CalendarItem calendarItem){
+        if (!securityUtil.isAdmin()) {
+            return "not authorized";
+        }
         calendarItemService.save(calendarItem);
         return "saved calendar item";
+
     }
 
     @PutMapping()
     public String updateCalendarItem(@RequestBody CalendarItem calendarItem){
+        if (securityUtil.isAdmin()) {
+            return "not authorized";
+        }
         calendarItemService.save(calendarItem);
         return "updated calendar item";
+
     }
     @DeleteMapping("/{id}")
     public String deleteCalendarItem(@PathVariable int id){
+        if (securityUtil.isAdmin()) {
+            return "not authorized";
+        }
         calendarItemService.deleteById(id);
         return "deleted calendar item";
+
     }
 }
